requirements.txt
astroid==2.5.6
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.8.2
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
